+++ packages/components/anchor-link/style/css.ts
@@ -1,0 +1,2 @@
+import '@element-plus/components/base/style/css'
+import '@element-plus/theme-chalk/el-anchor-link.css'
