+++ packages/components/anchor-link/style/index.ts
@@ -1,0 +1,2 @@
+import '@element-plus/components/base/style'
+import '@element-plus/theme-chalk/src/anchor-link.scss'
