+++ packages/theme-chalk/src/index.scss
@@ -105,3 +105,4 @@
 @use './statistic.scss';
 @use './tour.scss';
 @use './anchor.scss';
+@use './anchor-link.scss';
