+++ packages/utils/dom/index.ts
@@ -3,3 +3,4 @@ export * from './event'
 export * from './position'
 export * from './scroll'
 export * from './style'
+export * from './element'
