+++ packages/utils/dom/scroll.ts
@@ -1,4 +1,7 @@
 import { isClient } from '../browser'
+import { easeInOutCubic } from '../easings'
+import { isWindow } from '../types'
+import { rAF } from '../raf'
 import { getStyle } from './style'
 
 export const isScroll = (el: HTMLElement, isVertical?: boolean): boolean => {
@@ -98,4 +101,55 @@ export function scrollIntoView(
   } else if (bottom > viewRectBottom) {
     container.scrollTop = bottom - container.clientHeight
   }
+}
+
+export function animateScrollTo(
+  container: HTMLElement | Window,
+  from: number,
+  to: number,
+  duration: number,
+  callback?: unknown
+) {
+  const startTime = Date.now()
+
+  const scroll = () => {
+    const timestamp = Date.now()
+    const time = timestamp - startTime
+    const nextScrollTop = easeInOutCubic(
+      time > duration ? duration : time,
+      from,
+      to,
+      duration
+    )
+
+    if (isWindow(container)) {
+      container.scrollTo(window.pageXOffset, nextScrollTop)
+    } else {
+      container.scrollTop = nextScrollTop
+    }
+    if (time < duration) {
+      rAF(scroll)
+    } else if (typeof callback === 'function') {
+      callback()
+    }
+  }
+
+  scroll()
+}
+
+export const getScrollElement = (
+  target: HTMLElement,
+  container: HTMLElement | Window
+) => {
+  if (isWindow(container)) {
+    return target.ownerDocument.documentElement
+  }
+  return container
+}
+
+export const getScrollTop = (container: HTMLElement | Window) => {
+  if (isWindow(container)) {
+    return window.scrollY
+  }
+  return container.scrollTop
 }
