+++ packages/utils/index.ts
@@ -12,3 +12,5 @@ export * from './rand'
 export * from './strings'
 export * from './types'
 export * from './typescript'
+export * from './throttleByRaf'
+export * from './easings'
