+++ packages/utils/types.ts
@@ -36,3 +36,7 @@ export const isStringNumber = (val: string): boolean => {
   }
   return !Number.isNaN(Number(val))
 }
+
+export const isWindow = (val: unknown): val is Window => {
+  return val === window
+}
